producer: fix sFlow parsing when sampled header is truncated
We return early if we don't even have an Ethernet header. We also need
to increment the offset even if data is truncated to not confuse the
remaining of the packet with something else (eg extract TCP/UDP ports
when TCP/UDP header is too short).

Use IHL to correctly skip the IPv4 header when there are IP options.
Also fix the offset for TCP payload.

This also removes the encap/iteration loop. As encap is set to false at
the beginning of the loop, it was always executed one. I suppose this
was the remaining of some other code. The iterateMpls is removed by
moving some code around. One should look the diff with blanks ignored.